requirements.txt
pandas
cassandra-driver
prettytable